OAUTH2-131 Merge TestUtils into BaseTestPreparatorBundleActivator
