+++ modules/apps/oauth2-provider/oauth2-provider-test/src/main/java/com/liferay/oauth2/provider/test/internal/activator/BaseTestPreparatorBundleActivator.java
@@ -20,7 +20,7 @@ import com.liferay.oauth2.provider.scope.spi.prefix.handler.PrefixHandler;
 import com.liferay.oauth2.provider.scope.spi.prefix.handler.PrefixHandlerFactory;
 import com.liferay.oauth2.provider.scope.spi.scope.mapper.ScopeMapper;
 import com.liferay.oauth2.provider.service.OAuth2ApplicationLocalService;
-import com.liferay.oauth2.provider.test.internal.TestUtils;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.User;
@@ -28,14 +28,20 @@ import com.liferay.portal.kernel.service.CompanyLocalServiceUtil;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalServiceUtil;
 import com.liferay.portal.kernel.test.util.UserTestUtil;
+import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.io.IOException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Dictionary;
+import java.util.Enumeration;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Objects;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
@@ -43,9 +49,14 @@ import javax.ws.rs.core.Application;
 
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
+import org.osgi.framework.Constants;
+import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.framework.ServiceReference;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.cm.Configuration;
+import org.osgi.service.cm.ConfigurationAdmin;
+import org.osgi.service.cm.ManagedService;
+import org.osgi.service.cm.ManagedServiceFactory;
 import org.osgi.service.jaxrs.runtime.JaxrsServiceRuntime;
 import org.osgi.util.tracker.ServiceTracker;
 import org.osgi.util.tracker.ServiceTrackerCustomizer;
@@ -56,7 +67,38 @@ import org.osgi.util.tracker.ServiceTrackerCustomizer;
 public abstract class BaseTestPreparatorBundleActivator
 	implements BundleActivator {
 
-	public User addAdminUser(Company company) throws Exception {
+	@Override
+	public void start(BundleContext bundleContext) {
+		this.bundleContext = bundleContext;
+
+		autoCloseables = new ArrayList<>();
+
+		try {
+			prepareTest();
+		}
+		catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
+
+	@Override
+	public void stop(BundleContext bundleContext) {
+		ListIterator<AutoCloseable> listIterator = autoCloseables.listIterator(
+			autoCloseables.size());
+
+		while (listIterator.hasPrevious()) {
+			AutoCloseable previous = listIterator.previous();
+
+			try {
+				previous.close();
+			}
+			catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
+	}
+
+	protected User addAdminUser(Company company) throws Exception {
 		User user = UserTestUtil.addCompanyAdminUser(company);
 
 		autoCloseables.add(
@@ -65,7 +107,7 @@ public abstract class BaseTestPreparatorBundleActivator
 		return user;
 	}
 
-	public User addUser(Company company) throws Exception {
+	protected User addUser(Company company) throws Exception {
 		User user = UserTestUtil.addUser(company);
 
 		autoCloseables.add(
@@ -74,7 +116,33 @@ public abstract class BaseTestPreparatorBundleActivator
 		return user;
 	}
 
-	public Company createCompany(String hostName) throws PortalException {
+	protected Configuration configurationExists(
+			BundleContext bundleContext, String pid)
+		throws InvalidSyntaxException, IOException {
+
+		ServiceReference<ConfigurationAdmin> serviceReference =
+			bundleContext.getServiceReference(ConfigurationAdmin.class);
+
+		ConfigurationAdmin configurationAdmin = bundleContext.getService(
+			serviceReference);
+
+		try {
+			Configuration[] configurations =
+				configurationAdmin.listConfigurations(
+					"(" + Constants.SERVICE_PID + "=" + pid + ")");
+
+			if (ArrayUtil.isEmpty(configurations)) {
+				return null;
+			}
+
+			return configurations[0];
+		}
+		finally {
+			bundleContext.ungetService(serviceReference);
+		}
+	}
+
+	protected Company createCompany(String hostName) throws PortalException {
 		String virtualHostname = hostName + ".xyz";
 
 		Company company = CompanyLocalServiceUtil.addCompany(
@@ -87,7 +155,101 @@ public abstract class BaseTestPreparatorBundleActivator
 		return company;
 	}
 
-	public OAuth2Application createOAuth2Application(
+	protected Configuration createFactoryConfiguration(
+		BundleContext bundleContext, String factoryPid,
+		Dictionary<String, Object> properties) {
+
+		CountDownLatch countDownLatch = new CountDownLatch(1);
+
+		Dictionary<String, Object> registrationProperties =
+			new HashMapDictionary<>();
+
+		registrationProperties.put(Constants.SERVICE_PID, factoryPid);
+
+		ServiceRegistration<ManagedServiceFactory> serviceRegistration =
+			bundleContext.registerService(
+				ManagedServiceFactory.class,
+				new ManagedServiceFactory() {
+
+					@Override
+					public void deleted(String pid) {
+					}
+
+					@Override
+					public String getName() {
+						return
+							"Test managedservicefactory for pid " + factoryPid;
+					}
+
+					@Override
+					public void updated(
+						String pid, Dictionary<String, ?> updatedProperties) {
+
+						if (Validator.isNull(updatedProperties)) {
+							return;
+						}
+
+						if (isIncluded(properties, updatedProperties)) {
+							countDownLatch.countDown();
+						}
+					}
+
+				},
+				registrationProperties);
+
+		try {
+			ServiceReference<ConfigurationAdmin> serviceReference =
+				bundleContext.getServiceReference(ConfigurationAdmin.class);
+
+			ConfigurationAdmin configurationAdmin = bundleContext.getService(
+				serviceReference);
+
+			Configuration configuration = null;
+
+			try {
+				configuration = configurationAdmin.createFactoryConfiguration(
+					factoryPid, StringPool.QUESTION);
+
+				configuration.update(properties);
+
+				countDownLatch.await(5, TimeUnit.MINUTES);
+
+				return configuration;
+			}
+			catch (IOException ioe) {
+				throw new RuntimeException(ioe);
+			}
+			catch (InterruptedException ie) {
+				try {
+					configuration.delete();
+				}
+				catch (IOException ioe) {
+					throw new RuntimeException(ioe);
+				}
+
+				throw new RuntimeException(ie);
+			}
+			finally {
+				bundleContext.ungetService(serviceReference);
+			}
+		}
+		finally {
+			serviceRegistration.unregister();
+		}
+	}
+
+	protected Configuration createFactoryConfiguration(
+		String factoryPid, Dictionary<String, Object> properties) {
+
+		Configuration factoryConfiguration = createFactoryConfiguration(
+			bundleContext, factoryPid, properties);
+
+		autoCloseables.add(factoryConfiguration::delete);
+
+		return factoryConfiguration;
+	}
+
+	protected OAuth2Application createOAuth2Application(
 			long companyId, User user, String clientId)
 		throws PortalException {
 
@@ -99,7 +261,7 @@ public abstract class BaseTestPreparatorBundleActivator
 			Arrays.asList("everything", "everything.readonly"));
 	}
 
-	public OAuth2Application createOAuth2Application(
+	protected OAuth2Application createOAuth2Application(
 			long companyId, User user, String clientId,
 			List<GrantType> availableGrants, List<String> availableScopes)
 		throws PortalException {
@@ -109,7 +271,20 @@ public abstract class BaseTestPreparatorBundleActivator
 			availableGrants, availableScopes);
 	}
 
-	public OAuth2Application createOAuth2Application(
+	protected OAuth2Application createOAuth2Application(
+			long companyId, User user, String clientId,
+			List<String> availableScopes)
+		throws PortalException {
+
+		return createOAuth2Application(
+			companyId, user, clientId,
+			Arrays.asList(
+				GrantType.CLIENT_CREDENTIALS,
+				GrantType.RESOURCE_OWNER_PASSWORD),
+			availableScopes);
+	}
+
+	protected OAuth2Application createOAuth2Application(
 			long companyId, User user, String clientId, String clientSecret,
 			List<GrantType> availableGrants, List<String> availableScopes)
 		throws PortalException {
@@ -144,82 +319,52 @@ public abstract class BaseTestPreparatorBundleActivator
 		}
 	}
 
-	public ServiceRegistration<Application> registerJaxRsApplication(
-		Application application, String path,
-		Dictionary<String, Object> properties) {
+	protected void deleteConfiguration(
+		BundleContext bundleContext, String pid) {
 
-		if ((properties == null) || properties.isEmpty()) {
-			properties = new HashMapDictionary<>();
-		}
+		CountDownLatch countDownLatch = new CountDownLatch(1);
 
-		properties.put("oauth2.test.application", "true");
-		properties.put(
-			"osgi.jaxrs.extension.select", "(liferay.extension=OAuth2)");
-		properties.put("osgi.jaxrs.application.base", "/oauth2-test/" + path);
+		HashMapDictionary<String, Object> registrationProperties =
+			new HashMapDictionary<>();
 
-		ServiceRegistration<Application> serviceRegistration =
+		registrationProperties.put(Constants.SERVICE_PID, pid);
+
+		ServiceRegistration<ManagedService> serviceRegistration =
 			bundleContext.registerService(
-				Application.class, application, properties);
-
-		autoCloseables.add(serviceRegistration::unregister);
-
-		return serviceRegistration;
-	}
-
-	@Override
-	public void start(BundleContext bundleContext) {
-		this.bundleContext = bundleContext;
-
-		autoCloseables = new ArrayList<>();
+				ManagedService.class,
+				incomingProperties -> {
+					if (Validator.isNull(incomingProperties)) {
+						countDownLatch.countDown();
+					}
+				},
+				registrationProperties);
 
 		try {
-			prepareTest();
-		}
-		catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
+			ServiceReference<ConfigurationAdmin> serviceReference =
+				bundleContext.getServiceReference(ConfigurationAdmin.class);
 
-	@Override
-	public void stop(BundleContext bundleContext) {
-		ListIterator<AutoCloseable> listIterator = autoCloseables.listIterator(
-			autoCloseables.size());
-
-		while (listIterator.hasPrevious()) {
-			AutoCloseable previous = listIterator.previous();
+			ConfigurationAdmin configurationAdmin = bundleContext.getService(
+				serviceReference);
 
 			try {
-				previous.close();
+				Configuration configuration =
+					configurationAdmin.getConfiguration(
+						pid, StringPool.QUESTION);
+
+				configuration.delete();
+
+				countDownLatch.await(5, TimeUnit.MINUTES);
 			}
 			catch (Exception e) {
-				e.printStackTrace();
+				throw new RuntimeException(e);
+			}
+			finally {
+				bundleContext.ungetService(serviceReference);
 			}
 		}
-	}
-
-	protected Configuration createConfigurationFactory(
-		String factoryPid, Dictionary<String, Object> properties) {
-
-		Configuration factoryConfiguration =
-			TestUtils.createFactoryConfiguration(
-				bundleContext, factoryPid, properties);
-
-		autoCloseables.add(factoryConfiguration::delete);
-
-		return factoryConfiguration;
-	}
-
-	protected OAuth2Application createOAuth2Application(
-			long companyId, User user, String clientId,
-			List<String> availableScopes)
-		throws PortalException {
-
-		return createOAuth2Application(
-			companyId, user, clientId,
-			Arrays.asList(
-				GrantType.CLIENT_CREDENTIALS,
-				GrantType.RESOURCE_OWNER_PASSWORD),
-			availableScopes);
+		finally {
+			serviceRegistration.unregister();
+		}
 	}
 
 	protected void executeAndWaitForReadiness(Runnable runnable) {
@@ -268,7 +413,51 @@ public abstract class BaseTestPreparatorBundleActivator
 		}
 	}
 
+	protected boolean isIncluded(
+		Dictionary<String, ?> properties1, Dictionary<String, ?> properties2) {
+
+		if (properties1.size() > properties2.size()) {
+			return false;
+		}
+
+		Enumeration<String> keys = properties1.keys();
+
+		while (keys.hasMoreElements()) {
+			String key = keys.nextElement();
+
+			if (!Objects.deepEquals(
+					properties1.get(key), properties2.get(key))) {
+
+				return false;
+			}
+		}
+
+		return true;
+	}
+
 	protected abstract void prepareTest() throws Exception;
+
+	protected ServiceRegistration<Application> registerJaxRsApplication(
+		Application application, String path,
+		Dictionary<String, Object> properties) {
+
+		if ((properties == null) || properties.isEmpty()) {
+			properties = new HashMapDictionary<>();
+		}
+
+		properties.put("oauth2.test.application", "true");
+		properties.put(
+			"osgi.jaxrs.extension.select", "(liferay.extension=OAuth2)");
+		properties.put("osgi.jaxrs.application.base", "/oauth2-test/" + path);
+
+		ServiceRegistration<Application> serviceRegistration =
+			bundleContext.registerService(
+				Application.class, application, properties);
+
+		autoCloseables.add(serviceRegistration::unregister);
+
+		return serviceRegistration;
+	}
 
 	protected ServiceRegistration<PrefixHandlerFactory> registerPrefixHandler(
 		PrefixHandler prefixHandler, Dictionary<String, Object> properties) {
@@ -294,14 +483,79 @@ public abstract class BaseTestPreparatorBundleActivator
 		return serviceRegistration;
 	}
 
+	protected Configuration updateConfiguration(
+		BundleContext bundleContext, String servicePid,
+		Dictionary<String, ?> properties) {
+
+		CountDownLatch countDownLatch = new CountDownLatch(1);
+
+		HashMapDictionary<String, Object> registrationProperties =
+			new HashMapDictionary<>();
+
+		registrationProperties.put(Constants.SERVICE_PID, servicePid);
+
+		ServiceRegistration<ManagedService> serviceRegistration =
+			bundleContext.registerService(
+				ManagedService.class,
+				updatedProperties -> {
+					if (Validator.isNull(updatedProperties)) {
+						return;
+					}
+
+					if (isIncluded(properties, updatedProperties)) {
+						countDownLatch.countDown();
+					}
+				},
+				registrationProperties);
+
+		try {
+			ServiceReference<ConfigurationAdmin> serviceReference =
+				bundleContext.getServiceReference(ConfigurationAdmin.class);
+
+			ConfigurationAdmin configurationAdmin = bundleContext.getService(
+				serviceReference);
+
+			Configuration configuration = null;
+
+			try {
+				configuration = configurationAdmin.getConfiguration(
+					servicePid, StringPool.QUESTION);
+
+				configuration.update(properties);
+
+				countDownLatch.await(5, TimeUnit.MINUTES);
+
+				return configuration;
+			}
+			catch (IOException ioe) {
+				throw new RuntimeException(ioe);
+			}
+			catch (InterruptedException ie) {
+				try {
+					configuration.delete();
+				}
+				catch (IOException ioe) {
+					throw new RuntimeException(ioe);
+				}
+
+				throw new RuntimeException(ie);
+			}
+			finally {
+				bundleContext.ungetService(serviceReference);
+			}
+		}
+		finally {
+			serviceRegistration.unregister();
+		}
+	}
+
 	protected void updateOrCreateConfiguration(
 		String servicePid, Dictionary<String, ?> properties) {
 
 		Configuration configuration = null;
 
 		try {
-			configuration = TestUtils.configurationExists(
-				bundleContext, servicePid);
+			configuration = configurationExists(bundleContext, servicePid);
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -310,21 +564,19 @@ public abstract class BaseTestPreparatorBundleActivator
 		}
 
 		if (configuration == null) {
-			TestUtils.updateConfiguration(
-				bundleContext, servicePid, properties);
+			updateConfiguration(bundleContext, servicePid, properties);
 
 			autoCloseables.add(
-				() -> TestUtils.deleteConfiguration(bundleContext, servicePid));
+				() -> deleteConfiguration(bundleContext, servicePid));
 		}
 		else {
 			Dictionary<String, Object> oldProperties =
 				configuration.getProperties();
 
-			TestUtils.updateConfiguration(
-				bundleContext, servicePid, properties);
+			updateConfiguration(bundleContext, servicePid, properties);
 
 			autoCloseables.add(
-				() -> TestUtils.updateConfiguration(
+				() -> updateConfiguration(
 					bundleContext, servicePid, oldProperties));
 		}
 	}
